Set version to 0.3.5
Using version 1.0.0 is breaking "npm install appbuilder" and "npm install nativescript" on Windows where Visual Studio is not installed. Using version 1.0.0 is causing another dependency (ref-struct) to try using another version of ref (not this one) and trying to rebuild it. In case the machine do not have several building tools (python, Visual Studio 2010) this breaks the installation.